pyproject.toml
[build-system]
requires = ["setuptools>=68.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "autoguard-ml"
version = "0.1.0"
description = "Unified AutoML + Dataset Diagnosis + Drift Detection for production ML"
readme = "README.md"
license = { file = "LICENSE" }
authors = [{ name = "AutoGuard Contributors", email = "hello@autoguard-ml.dev" }]
keywords = ["automl", "drift-detection", "mlops", "explainability", "data-quality", "machine-learning"]
classifiers = [
    "Development Status :: 4 - Beta",
    "Intended Audience :: Developers",
    "Intended Audience :: Science/Research",
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Topic :: Scientific/Engineering :: Artificial Intelligence",
]
requires-python = ">=3.10"
dependencies = [
    "numpy>=1.24.0",
    "pandas>=2.0.0",
    "scikit-learn>=1.3.0",
    "xgboost>=2.0.0",
    "lightgbm>=4.0.0",
    "optuna>=3.4.0",
    "shap>=0.43.0",
    "matplotlib>=3.7.0",
    "seaborn>=0.12.0",
    "scipy>=1.11.0",
    "pyyaml>=6.0",
    "rich>=13.0.0",
    "click>=8.1.0",
    "joblib>=1.3.0",
    "imbalanced-learn>=0.11.0",
    "tabulate>=0.9.0",
    "tqdm>=4.66.0",
    "jinja2>=3.1.0",
]

[project.optional-dependencies]
api    = ["fastapi>=0.104.0", "uvicorn[standard]>=0.24.0", "pydantic>=2.0.0"]
dashboard = ["streamlit>=1.28.0", "plotly>=5.18.0"]
dev    = ["pytest>=7.4.0", "pytest-cov>=4.1.0", "black>=23.0.0", "ruff>=0.1.0"]
all    = ["autoguard-ml[api,dashboard,dev]"]

[project.scripts]
autoguard = "autoguard.cli.main:cli"

[project.urls]
Homepage = "https://github.com/autoguard/autoguard-ml"
Documentation = "https://autoguard-ml.readthedocs.io"

[tool.setuptools.packages.find]
where = ["."]
include = ["autoguard*"]

[tool.pytest.ini_options]
testpaths = ["tests"]
addopts = "-v"
